+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Change `RSpec/ContainExactly` to ignore calls with no arguments, and change `RSpec/MatchArray` to ignore calls with an empty array literal argument. ([@ydah], [@bquorning])
 - Add new `RSpec/BeEmpty` cop. ([@ydah], [@bquorning])
 - Make `RSpec/MatchArray` and `RSpec/ContainExactly` pending. ([@ydah])
+- Add autocorrect support for `RSpec/ScatteredSetup`. ([@ydah])
 
 ## 2.19.0 (2023-03-06)